企业版提取代码

+++ internal/model/entity/crontab_base.go
@@ -25,8 +25,8 @@ type CrontabBase struct {
 	CrontabId          uint   `json:"crontab_id"            ` // 任务编号
 	CrontabName        string `json:"crontab_name"          ` // 任务名称
 	CrontabFile        string `json:"crontab_file"          ` // 任务脚本
-	CrontabLastExeTime uint   `json:"crontab_last_exe_time" ` // 上次执行时间
-	CrontabNextExeTime uint   `json:"crontab_next_exe_time" ` // 下次执行时间
+	CrontabLastExeTime uint64 `json:"crontab_last_exe_time" ` // 上次执行时间
+	CrontabNextExeTime uint64 `json:"crontab_next_exe_time" ` // 下次执行时间
 	CrontabMinute      string `json:"crontab_minute"        ` // 分钟(LIST):*-每分; 0-0;1-1; 2-2; 3-3; 4-4; 5-5; 6-6; 7-7; 8-8; 9-9; 10-10; 11-11; 12-12; 13-13; 14-14; 15-15; 16-16; 17-17; 18-18; 19-19; 20-20; 21-21; 22-22; 23-23; 24-24; 25-25; 26-26; 27-27; 28-28; 29-29; 30-30; 31-31; 32-32; 33-33; 34-34; 35-35; 36-36; 37-37; 38-38; 39-39; 40-40; 41-41; 42-42; 43-43; 44-44; 45-45; 46-46; 47-47; 48-48; 49-49; 50-50; 51-51; 52-52; 53-53; 54-54; 55-55; 56-56; 57-57; 58-58; 59-59
 	CrontabHour        string `json:"crontab_hour"          ` // 小时(LIST):*-任意; 0-0;1-1; 2-2; 3-3; 4-4; 5-5; 6-6; 7-7; 8-8; 9-9; 10-10; 11-11; 12-12; 13-13; 14-14; 15-15; 16-16; 17-17; 18-18; 19-19; 20-20; 21-21; 22-22; 23-23
 	CrontabDay         string `json:"crontab_day"           ` // 每天(LIST):*-任意;1-1; 2-2; 3-3; 4-4; 5-5; 6-6; 7-7; 8-8; 9-9; 10-10; 11-11; 12-12; 13-13; 14-14; 15-15; 16-16; 17-17; 18-18; 19-19; 20-20; 21-21; 22-22; 23-23; 24-24; 25-25; 26-26; 27-27; 28-28; 29-29; 30-30; 31-31
